Move MDX compiler to csf-tools
