--- addons/docs/src/mdx/mdx-compiler-plugin.ts
@@ -1,1 +1,0 @@
-export { createCompiler } from '@storybook/mdx';
