+++ lib/csf-tools/README.md
@@ -1,10 +1,12 @@
 # Storybook CSF Tools
 
-A library to read, analyze, transform, and write CSF programmatically.
+An experimental library to read, analyze, transform, and write CSF programmatically.
 
 - Read - Parse a CSF file with Babel
 - Analyze - Extract its metadata & stories based on the Babel AST
 - Write - Write the AST back to a file
+
+It can can parse MDX into CSF.
 
 Coming soon:
