+++ lib/csf-tools/mdx.d.ts
@@ -1,0 +1,1 @@
+export * from './dist/ts3.9/mdx/index.d';
