+++ lib/csf-tools/mdx.js
@@ -1,0 +1,1 @@
+module.exports = require('./dist/cjs/mdx');
