+++ lib/csf-tools/package.json
@@ -42,16 +42,23 @@
   "dependencies": {
     "@babel/generator": "^7.12.11",
     "@babel/parser": "^7.12.11",
+    "@babel/plugin-transform-react-jsx": "^7.12.12",
+    "@babel/preset-env": "^7.12.11",
     "@babel/traverse": "^7.12.11",
     "@babel/types": "^7.12.11",
+    "@mdx-js/mdx": "^1.6.22",
     "@storybook/csf": "^0.0.1",
     "core-js": "^3.8.2",
     "fs-extra": "^9.0.1",
+    "js-string-escape": "^1.0.1",
+    "lodash": "^4.17.20",
+    "prettier": "~2.2.1",
     "regenerator-runtime": "^0.13.7"
   },
   "devDependencies": {
     "@types/fs-extra": "^9.0.6",
-    "js-yaml": "^3.14.1"
+    "js-yaml": "^3.14.1",
+    "ts-dedent": "^2.0.0"
   },
   "publishConfig": {
     "access": "public"
