+++ lib/csf-tools/src/__testfixtures__/component-args.output.snapshot
