+++ lib/csf-tools/src/__testfixtures__/component-id.output.snapshot
