+++ lib/csf-tools/src/__testfixtures__/decorators.output.snapshot
