+++ lib/csf-tools/src/__testfixtures__/non-story-exports.output.snapshot
