+++ lib/csf-tools/src/__testfixtures__/story-args.output.snapshot
