+++ lib/csf-tools/src/__testfixtures__/story-current.output.snapshot
