+++ lib/csf-tools/src/__testfixtures__/story-def-text-only.output.snapshot
