+++ lib/csf-tools/src/__testfixtures__/story-definitions.output.snapshot
