+++ lib/csf-tools/src/__testfixtures__/story-function-var.output.snapshot
