+++ lib/csf-tools/src/__testfixtures__/story-function.output.snapshot
