+++ lib/csf-tools/src/__testfixtures__/story-multiple-children.output.snapshot
