+++ lib/csf-tools/src/__testfixtures__/story-object.output.snapshot
