+++ lib/csf-tools/src/__testfixtures__/vanilla.output.snapshot
