--- lib/mdx/README.md
@@ -1,3 +1,0 @@
-# Storybook MDX
-
-Storybook MDX is an internal utility package to parse MDX files and compile them into Storybook's Component Story Format. For more information, see [Storybook Docs](https://storybook.js.org/docs/react/writing-docs/introduction) documentation.
