--- lib/mdx/package.json
@@ -1,65 +1,0 @@
-{
-  "name": "@storybook/mdx",
-  "version": "6.3.0-rc.2",
-  "description": "Storybook MDX utilities",
-  "keywords": [
-    "mdx",
-    "storybook"
-  ],
-  "homepage": "https://github.com/storybookjs/storybook/tree/master/addons/docs",
-  "bugs": {
-    "url": "https://github.com/storybookjs/storybook/issues"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/storybookjs/storybook.git",
-    "directory": "lib/mdex"
-  },
-  "funding": {
-    "type": "opencollective",
-    "url": "https://opencollective.com/storybook"
-  },
-  "license": "MIT",
-  "main": "dist/cjs/index.js",
-  "module": "dist/esm/index.js",
-  "types": "dist/ts3.9/index.d.ts",
-  "typesVersions": {
-    "<3.8": {
-      "*": [
-        "dist/ts3.4/*"
-      ]
-    }
-  },
-  "files": [
-    "dist/**/*",
-    "README.md",
-    "*.js",
-    "*.d.ts",
-    "!__testfixtures__"
-  ],
-  "scripts": {
-    "prepare": "node ../../scripts/prepare.js"
-  },
-  "dependencies": {
-    "@babel/generator": "^7.12.11",
-    "@babel/parser": "^7.12.11",
-    "@babel/plugin-transform-react-jsx": "^7.12.12",
-    "@babel/preset-env": "^7.12.11",
-    "@babel/types": "^7.12.11",
-    "@mdx-js/mdx": "^1.6.22",
-    "js-string-escape": "^1.0.1",
-    "lodash": "^4.17.20",
-    "prettier": "~2.2.1",
-    "regenerator-runtime": "^0.13.7",
-    "ts-dedent": "^2.0.0"
-  },
-  "devDependencies": {
-    "jest": "^26.6.3",
-    "jest-specific-snapshot": "^4.0.0"
-  },
-  "publishConfig": {
-    "access": "public"
-  },
-  "gitHead": "0aaad2c19224858494bca5e712b3029cd66a2f6c",
-  "sbmodern": "dist/modern/index.js"
-}
