+++ nx.json
@@ -76,9 +76,6 @@
     "@storybook/preact": {
       "implicitDependencies": []
     },
-    "@storybook/mdx": {
-      "implicitDependencies": []
-    },
     "@storybook/react": {
       "implicitDependencies": []
     },
