+++ package.json
@@ -157,7 +157,6 @@
     "@storybook/eslint-config-storybook": "^2.4.0",
     "@storybook/html": "workspace:*",
     "@storybook/linter-config": "^2.5.0",
-    "@storybook/mdx": "workspace:*",
     "@storybook/node-logger": "workspace:*",
     "@storybook/postinstall": "workspace:*",
     "@storybook/preact": "workspace:*",
