+++ workspace.json
@@ -73,10 +73,6 @@
       "root": "lib/manager-webpack5",
       "type": "library"
     },
-    "@storybook/mdx": {
-      "root": "lib/mdx",
-      "type": "library"
-    },
     "@storybook/preact": {
       "root": "app/preact",
       "type": "library"
